reconstruct file arrangement

+++ remove_and_create_dirs.sh
@@ -1,24 +1,33 @@
 #!/bin/bash
 
-source ./config.sh
-
 create_if_nonexist()
 {
   for dir in "$@"; do
    if [ ! -d "$dir" ]; then
+     echo "Creating $dir..."
      mkdir "$dir"
    fi
   done
 }
 
-create_if_nonexist "$MY_WORKSPACE" "$MY_DOWNLOADS"
-
 del_dirs()
 {
   for dir in "$@"; do
-    rm -rf "$dir"
+    if [ -d "$dir" ]; then
+      echo "Deleting $dir..."
+      rm -rf "$dir"
+    fi
   done
 }
+
+source ./config.sh
+
+config_dir="$MY_HOME/.config"
+wanted_dirs=($MY_WORKSPACE $MY_DOWNLOADS $config_dir)
+
+for dir_name in ${wanted_dirs[@]}; do
+  create_if_nonexist $dir_name
+done
 
 documents="$MY_HOME/Documents"
 downloads="$MY_HOME/Downloads"
